Finalizing the lit-shader refactoring by moving shader code to chunks
