+++ src/scene/shader-lib/programs/lit-shader.js
@@ -4,10 +4,9 @@ import {
     SEMANTIC_TEXCOORD0, SEMANTIC_TEXCOORD1
 } from '../../../platform/graphics/constants.js';
 import {
-    LIGHTSHAPE_PUNCTUAL, LIGHTSHAPE_RECT, LIGHTSHAPE_DISK, LIGHTSHAPE_SPHERE,
+    LIGHTSHAPE_PUNCTUAL,
     LIGHTTYPE_DIRECTIONAL, LIGHTTYPE_OMNI, LIGHTTYPE_SPOT,
     SHADER_DEPTH, SHADER_PICK,
-    SHADOW_PCF3_32F, SHADOW_VSM_32F,
     SPRITE_RENDERMODE_SLICED, SPRITE_RENDERMODE_TILED, shadowTypeInfo, SHADER_PREPASS,
     lightTypeNames, lightShapeNames, spriteRenderModeNames, fresnelNames, blendNames, lightFalloffNames,
     cubemaProjectionNames, specularOcclusionNames, reflectionSrcNames, ambientSrcNames,
@@ -17,7 +16,6 @@ import { shaderChunks } from '../chunks/chunks.js';
 import { ChunkUtils } from '../chunk-utils.js';
 import { ShaderPass } from '../../shader-pass.js';
 import { validateUserChunks } from '../chunks/chunk-validation.js';
-import { ChunkBuilder } from '../chunk-builder.js';
 import { Debug } from '../../../core/debug.js';
 
 /**
@@ -49,6 +47,20 @@ class LitShader {
      * @type {string}
      */
     varyingsCode = '';
+
+    /**
+     * The graphics device.
+     *
+     * @type {GraphicsDevice}
+     */
+    device;
+
+    /**
+     * The lit options.
+     *
+     * @type {LitShaderOptions}
+     */
+    options;
 
     /**
      * @param {GraphicsDevice} device - The graphics device.
@@ -114,14 +126,7 @@ class LitShader {
         // generated by vshader
         this.vshader = null;
 
-        // supplied by caller
-        this.frontendDecl = null;
-        this.frontendCode = null;
-        this.lightingUv = null;
-
         // defines set by the shader generation
-        this.defines = [];
-
         this.vDefines = new Map();
         this.fDefines = new Map();
 
@@ -129,22 +134,23 @@ class LitShader {
         this.fshader = null;
     }
 
-    _getLightSourceShapeString(shape) {
-        switch (shape) {
-            case LIGHTSHAPE_RECT:
-                return 'Rect';
-            case LIGHTSHAPE_DISK:
-                return 'Disk';
-            case LIGHTSHAPE_SPHERE:
-                return 'Sphere';
-            default:
-                return '';
+    /**
+     * Helper function to define a value in the fragment shader.
+     *
+     * @param {boolean} condition - The define is added if the condition is true.
+     * @param {string} name - The define name.
+     * @param {string} [value] - The define value.
+     */
+    fDefineSet(condition, name, value = '') {
+        if (condition) {
+            this.fDefines.set(name, value);
         }
     }
 
     /**
      * The function generates defines for the lit vertex shader, and handles required attributes and
-     * varyings. The source code of the shader is supplied by litMainVS chunk.
+     * varyings. The source code of the shader is supplied by litMainVS chunk. This vertex shader is
+     * used for all render passes.
      *
      * @param {any} useUv - Info about used UVs.
      * @param {any} useUnmodifiedUv - Info about used unmodified UVs.
@@ -290,110 +296,6 @@ class LitShader {
         this.vshader = this.varyingsCode + chunks.litMainVS;
     }
 
-    _fsGetBeginCode() {
-
-        let code = '';
-        for (let i = 0; i < this.defines.length; i++) {
-            code += `#define ${this.defines[i]}\n`;
-        }
-
-        return code;
-    }
-
-    _fsGetShadowPassCode() {
-        const options = this.options;
-
-        const lightType = this.shaderPassInfo.lightType;
-        let shadowType = this.shaderPassInfo.shadowType;
-
-        const shadowInfo = shadowTypeInfo.get(shadowType);
-        Debug.assert(shadowInfo);
-        const { vsm, pcss, pcf } = shadowInfo;
-
-        // If not a directional light and using clustered, fall back to using PCF3x3 if shadow type isn't supported
-        if (lightType !== LIGHTTYPE_DIRECTIONAL && options.clusteredLightingEnabled) {
-            if (!pcf) {
-                shadowType = SHADOW_PCF3_32F;
-            }
-        }
-
-        // Use perspective depth for:
-        // Directional: Always since light has no position
-        // Spot: If not using VSM
-        // Point: Never
-        const usePerspectiveDepth = (lightType === LIGHTTYPE_DIRECTIONAL || (!vsm && lightType === LIGHTTYPE_SPOT));
-
-        const code = `
-
-            ${this._fsGetBeginCode()}
-            ${this.varyingsCode}
-            ${this.frontendDecl}
-            ${this.frontendCode}
-
-            // Flag if we are using non-standard depth, i.e gl_FragCoord.z
-            ${usePerspectiveDepth ? '#define PERSPECTIVE_DEPTH' : ''}
-            ${pcss ? '#define PCSS' : ''}
-            ${vsm ? '#define VSM' : ''}
-            ${shadowType === SHADOW_VSM_32F ? '#define VSM_32F' : ''}
-            ${lightType === LIGHTTYPE_DIRECTIONAL ? '#define DIRECTIONAL_LIGHT' : ''}
-
-            #ifndef DIRECTIONAL_LIGHT
-                uniform vec3 view_position;
-                uniform float light_radius;
-            #endif
-
-            #ifdef PCSS
-                #include "linearizeDepthPS"
-            #endif
-
-            void main(void) {
-                evaluateFrontend();
-
-                #ifdef PERSPECTIVE_DEPTH
-                    float depth = gl_FragCoord.z;
-
-                    #ifdef PCSS
-                        // spot/omni shadows currently use linear depth.
-                        // TODO: use perspective depth for spot/omni the same way as directional
-                        #ifndef DIRECTIONAL_LIGHT
-                            depth = linearizeDepth(depth, camera_params);
-                        #endif
-                    #endif
-
-                #else
-                    float depth = min(distance(view_position, vPositionW) / light_radius, 0.99999);
-                    #define MODIFIED_DEPTH
-                #endif
-
-                #ifdef VSM
-                    #ifdef VSM_32F
-                        float exponent = 15.0;
-                    #else
-                        float exponent = 5.54;
-                    #endif
-                    depth = 2.0 * depth - 1.0;
-                    depth =  exp(exponent * depth);
-                    gl_FragColor = vec4(depth, depth*depth, 1.0, 1.0);
-                #else
-                    #ifdef PCSS
-                        // store depth into R32
-                        gl_FragColor.r = depth;
-                    #else
-                        #ifdef MODIFIED_DEPTH
-                            // If we end up using modified depth, it needs to be explicitly written to gl_FragDepth
-                            gl_FragDepth = depth;
-                        #endif
-
-                        // just the simplest code, color is not written anyway
-                        gl_FragColor = vec4(1.0);
-                    #endif
-                #endif
-            }
-        `;
-
-        return code;
-    }
-
     /**
      * Generate defines for lighting environment as well as individual lights.
      *
@@ -483,21 +385,8 @@ class LitShader {
         }
     }
 
-    fDefineSet(condition, name, value = '') {
-        if (condition) {
-            this.fDefines.set(name, value);
-        }
-    }
-
-    // the big one
-    _fsGetLitPassCode() {
+    prepareForwardPass(lightingUv) {
         const { options } = this;
-
-        const decl = new ChunkBuilder();
-        const func = new ChunkBuilder();
-        const backend = new ChunkBuilder();
-
-        const hasTBN = this.needsNormal && (options.useNormals || options.useClearCoatNormals || (options.enableGGXSpecular && !options.useHeights));
 
         // area lights are used when clustered area lights are enabled or any lights have area shape
         const clusteredAreaLights = options.clusteredLightingEnabled && options.clusteredLightingAreaLightsEnabled;
@@ -505,6 +394,7 @@ class LitShader {
             return light._shape && light._shape !== LIGHTSHAPE_PUNCTUAL;
         });
         const addAmbient = !options.lightMapEnabled || options.lightMapWithoutAmbient;
+        const hasTBN = this.needsNormal && (options.useNormals || options.useClearCoatNormals || (options.enableGGXSpecular && !options.useHeights));
 
         if (options.useSpecular) {
             this.fDefineSet(true, 'LIT_SPECULAR');
@@ -526,7 +416,7 @@ class LitShader {
         this.fDefineSet(options.twoSidedLighting, 'LIT_TWO_SIDED_LIGHTING');
         this.fDefineSet(options.lightMapEnabled, 'LIT_LIGHTMAP');
         this.fDefineSet(options.dirLightMapEnabled, 'LIT_DIR_LIGHTMAP');
-        this.fDefineSet(options.skyboxIntensity, 'LIT_SKYBOX_INTENSITY');
+        this.fDefineSet(options.skyboxIntensity > 0, 'LIT_SKYBOX_INTENSITY');
         this.fDefineSet(options.clusteredLightingShadowsEnabled, 'LIT_CLUSTERED_SHADOWS');
         this.fDefineSet(options.clusteredLightingAreaLightsEnabled, 'LIT_CLUSTERED_AREA_LIGHTS');
         this.fDefineSet(hasTBN, 'LIT_TBN');
@@ -555,525 +445,32 @@ class LitShader {
         this.fDefineSet(true, 'LIT_AMBIENT_SOURCE', ambientSrcNames[options.ambientSource]);
 
         // injection defines
-        this.fDefineSet(true, '{lightingUv}', this.lightingUv ?? ''); // example: vUV0_1
+        this.fDefineSet(true, '{lightingUv}', lightingUv ?? ''); // example: vUV0_1
         this.fDefineSet(true, '{reflectionDecode}', ChunkUtils.decodeFunc(options.reflectionEncoding));
         this.fDefineSet(true, '{reflectionCubemapDecode}', ChunkUtils.decodeFunc(options.reflectionCubemapEncoding));
         this.fDefineSet(true, '{ambientDecode}', ChunkUtils.decodeFunc(options.ambientEncoding));
 
         // lighting defines
         this._setupLightingDefines(hasAreaLights, options.clusteredLightingEnabled);
+    }
 
-        decl.append(`
-            // globals
-            vec3 sReflection;
-            vec3 dVertexNormalW;
-            vec3 dTangentW;
-            vec3 dBinormalW;
-            vec3 dViewDirW;
-            vec3 dReflDirW;
-            vec3 dHalfDirW;
-            vec3 ccReflDirW;
+    prepareShadowPass() {
 
-            // Per-light temporaries
-            vec3 dLightDirNormW;
-            vec3 dLightDirW;
-            vec3 dLightPosW;
-            vec3 dShadowCoord;
+        const lightType = this.shaderPassInfo.lightType;
 
-            // Outputs
-            mat3 dTBN;
-            vec4 dReflection;
-            vec3 dDiffuseLight;
-            vec3 dSpecularLight;
-            float dAtten;
-            float dAttenD; // separate diffuse attenuation for non-punctual light sources
-            vec3 dAtten3;
-            vec4 dMsdf;
-            float ccFresnel;
-            vec3 ccReflection;
-            vec3 ccSpecularLight;
-            float ccSpecularityNoFres;
-            vec3 sSpecularLight;
+        const shadowType = this.shaderPassInfo.shadowType;
+        const shadowInfo = shadowTypeInfo.get(shadowType);
+        Debug.assert(shadowInfo);
 
-            // FRAGMENT SHADER INPUTS: UNIFORMS
+        // Use perspective depth for:
+        // - Directional: Always since light has no position
+        // - Spot: If not using VSM
+        // - Point: Never
+        const usePerspectiveDepth = (lightType === LIGHTTYPE_DIRECTIONAL || (!shadowInfo.vsm && lightType === LIGHTTYPE_SPOT));
 
-            #ifdef LIT_DISPERSION
-                uniform float material_dispersion;
-            #endif
-
-            #ifndef LIT_OPACITY_FADES_SPECULAR
-                uniform float material_alphaFade;
-            #endif
-
-            #ifdef LIT_SSAO
-                uniform sampler2D ssaoTexture;
-                uniform vec2 ssaoTextureSizeInv;
-            #endif
-
-            // lighting and shadowing declarations
-
-            // LOOP - uniform declarations for all non-clustered lights
-            #include "lightDeclarationPS, LIGHT_COUNT"
-
-            #ifdef LIT_SPECULAR
-                #if LIT_FRESNEL_MODEL == NONE && !defined(LIT_REFLECTIONS) && !defined(LIT_DIFFUSE_MAP) 
-                    #define LIT_OLD_AMBIENT
-                #endif
-            #endif
-
-            // TODO: move this to 'func' section, but for now we prepend to it and it needs to be first
-            #include "basePS"
-
-        `);
-
-        func.append(`
-
-            // 9-slice support code
-            #if LIT_NONE_SLICE_MODE == SLICED
-                #include "baseNineSlicedPS"
-            #elif LIT_NONE_SLICE_MODE == TILED
-                #include "baseNineSlicedTiledPS"
-            #endif
-
-            // TBN
-            #ifdef LIT_TBN
-                #include "TBNPS"
-
-                #ifdef LIT_TWO_SIDED_LIGHTING
-                    #include "twoSidedLightingPS"
-                #endif
-            #endif
-
-            #include "sphericalPS"
-            #include "decodePS"
-            #include "gammaPS"
-            #include "tonemappingPS"
-            #include "fogPS"
-        `);
-
-        // frontend
-        func.append(this.frontendCode);
-
-        func.append(`
-            #ifdef LIT_NEEDS_NORMAL
-                #include "cubeMapRotatePS"
-                #include "cubeMapProjectPS"
-                #include "envProcPS"
-            #endif
-
-            // ----- specular or reflections -----
-            #if defined(LIT_LIGHTING) && defined(LIT_SPECULAR)
-                #define LIT_SPECULAR_OR_REFLECTION
-            #elif defined(LIT_REFLECTIONS)
-                #define LIT_SPECULAR_OR_REFLECTION
-            #endif
-
-            #ifdef LIT_SPECULAR_OR_REFLECTION
-
-                #ifdef LIT_METALNESS
-                    #include "metalnessModulatePS"
-                #endif
-
-                #if LIT_FRESNEL_MODEL == SCHLICK
-                    #include "fresnelSchlickPS"
-                #endif
-
-                #ifdef LIT_IRIDESCENCE
-                    #include "iridescenceDiffractionPS"
-                #endif
-            #endif
-
-            // ----- ambient occlusion -----
-            #ifdef LIT_AO
-                #include "aoDiffuseOccPS"
-                #include "aoSpecOccPS"
-            #endif
-
-            #if LIT_REFLECTION_SOURCE == ENVATLASHQ
-                #include "envAtlasPS"
-                #include "reflectionEnvHQPS"
-            #elif LIT_REFLECTION_SOURCE == ENVATLAS
-                #include "envAtlasPS"
-                #include "reflectionEnvPS"
-            #elif LIT_REFLECTION_SOURCE == CUBEMAP
-                #include "reflectionCubePS"
-            #elif LIT_REFLECTION_SOURCE == SPHEREMAP
-                #include "reflectionSpherePS"
-            #endif
-
-            #ifdef LIT_REFLECTIONS
-                #ifdef LIT_CLEARCOAT
-                    #include "reflectionCCPS"
-                #endif
-
-                #ifdef LIT_SHEEN
-                    #include "reflectionSheenPS"
-                #endif
-            #endif
-
-            #ifdef LIT_REFRACTION
-                #if defined(LIT_DYNAMIC_REFRACTION)
-                    #include "refractionDynamicPS"
-                #elif defined(LIT_REFLECTIONS)
-                    #include "refractionCubePS"
-                #endif
-            #endif
-
-            #ifdef LIT_SHEEN
-                #include "lightSheenPS"
-            #endif
-
-            #ifdef LIT_GGX_SPECULAR
-                uniform float material_anisotropy;
-            #endif
-
-            uniform vec3 material_ambient;
-
-            #ifdef LIT_SPECULAR
-                #ifdef LIT_LIGHTING
-                    #ifdef LIT_GGX_SPECULAR
-                        #include "lightSpecularAnisoGGXPS"
-                    #else
-                        #include "lightSpecularBlinnPS"
-                    #endif
-                #endif
-            #endif
-
-            #include "combinePS"
-
-            #ifdef LIT_LIGHTMAP
-                #include "lightmapAddPS"
-            #endif
-
-            #ifdef LIT_ADD_AMBIENT
-                #include "ambientPS"
-            #endif
-
-            #ifdef LIT_MSDF
-                #include "msdfPS"
-            #endif
-
-            #ifdef LIT_NEEDS_NORMAL
-                #include "viewDirPS"
-                #ifdef LIT_SPECULAR
-                    #ifdef LIT_GGX_SPECULAR
-                        #include "reflDirAnisoPS"
-                    #else
-                        #include "reflDirPS"
-                    #endif
-                #endif
-            #endif
-
-            // required lighting functionality
-            #include "lightingPS"
-
-        `);
-
-        // FRAGMENT SHADER BODY
-
-        const code = `
-
-            void main(void) {
-                dReflection = vec4(0);
-
-                #ifdef LIT_CLEARCOAT
-                    ccSpecularLight = vec3(0);
-                    ccReflection = vec3(0);
-                #endif
-
-                #if LIT_NONE_SLICE_MODE == SLICED
-                    #include "startNineSlicedPS"
-                #elif LIT_NONE_SLICE_MODE == TILED
-                    #include "startNineSlicedTiledPS"
-                #endif
-
-                #ifdef LIT_NEEDS_NORMAL
-                    dVertexNormalW = normalize(vNormalW);
-
-                    #ifdef LIT_TANGENTS
-                        #if defined(LIT_HEIGHTS) || defined(LIT_USE_NORMALS)
-                            dTangentW = vTangentW;
-                            dBinormalW = vBinormalW;
-                        #endif
-                    #endif
-
-                    getViewDir();
-
-                    #ifdef LIT_TBN
-                        getTBN(dTangentW, dBinormalW, dVertexNormalW);
-
-                        #ifdef LIT_TWO_SIDED_LIGHTING
-                            handleTwoSidedLighting();
-                        #endif
-                    #endif
-                #endif
-
-                // invoke frontend functions
-                evaluateFrontend();
-
-                #include "debugProcessFrontendPS"
-
-                evaluateBackend();
-            }
-        `;
-
-        backend.append(`
-
-            // apply SSAO during lighting
-            #ifdef LIT_SSAO
-                litArgs_ao *= texture2DLod(ssaoTexture, gl_FragCoord.xy * ssaoTextureSizeInv, 0.0).r;
-            #endif
-
-            // transform tangent space normals to world space
-            #ifdef LIT_NEEDS_NORMAL
-                #ifdef LIT_SPECULAR
-                    getReflDir(litArgs_worldNormal, dViewDirW, litArgs_gloss, dTBN);
-                #endif
-
-                #ifdef LIT_CLEARCOAT
-                    ccReflDirW = normalize(-reflect(dViewDirW, litArgs_clearcoat_worldNormal));
-                #endif
-            #endif
-
-            #ifdef LIT_SPECULAR_OR_REFLECTION
-                #ifdef LIT_METALNESS
-                    float f0 = 1.0 / litArgs_ior; f0 = (f0 - 1.0) / (f0 + 1.0); f0 *= f0;
-                    litArgs_specularity = getSpecularModulate(litArgs_specularity, litArgs_albedo, litArgs_metalness, f0);
-                    litArgs_albedo = getAlbedoModulate(litArgs_albedo, litArgs_metalness);
-                #endif
-
-                #ifdef LIT_IRIDESCENCE
-                    vec3 iridescenceFresnel = getIridescence(saturate(dot(dViewDirW, litArgs_worldNormal)), litArgs_specularity, litArgs_iridescence_thickness);
-                #endif
-            #endif
-
-            // ambient
-            #ifdef LIT_ADD_AMBIENT
-                addAmbient(litArgs_worldNormal);
-
-                #ifdef LIT_SPECULAR
-                    dDiffuseLight = dDiffuseLight * (1.0 - litArgs_specularity);
-                #endif
-
-                // move ambient color out of diffuse (used by Lightmapper, to multiply ambient color by accumulated AO)
-                #ifdef LIT_SEPARATE_AMBIENT
-                    vec3 dAmbientLight = dDiffuseLight;
-                    dDiffuseLight = vec3(0);
-                #endif
-            #endif
-
-            #ifndef LIT_OLD_AMBIENT
-                dDiffuseLight *= material_ambient;
-            #endif
-
-            #ifdef LIT_AO
-                #ifndef LIT_OCCLUDE_DIRECT
-                    occludeDiffuse(litArgs_ao);
-                #endif
-            #endif
-
-            #ifdef LIT_LIGHTMAP
-                addLightMap(
-                    litArgs_lightmap, 
-                    litArgs_lightmapDir, 
-                    litArgs_worldNormal, 
-                    dViewDirW, 
-                    dReflDirW, 
-                    litArgs_gloss, 
-                    litArgs_specularity, 
-                    dVertexNormalW,
-                    dTBN
-                #if defined(LIT_IRIDESCENCE)
-                    , iridescenceFresnel,
-                    litArgs_iridescence_intensity
-                #endif
-                );
-            #endif
-        `);
-
-        if (this.lighting || this.reflections) {
-            backend.append(`
-                
-                #ifdef LIT_REFLECTIONS
-
-                    #ifdef LIT_CLEARCOAT
-                        addReflectionCC(ccReflDirW, litArgs_clearcoat_gloss);
-                    
-                        #ifdef LIT_SPECULAR_FRESNEL
-                            ccFresnel = getFresnelCC(dot(dViewDirW, litArgs_clearcoat_worldNormal));
-                            ccReflection.rgb *= ccFresnel;
-                        #else
-                            ccFresnel = 0.0;
-                        #endif
-                    #endif
-
-                    #ifdef LIT_SPECULARITY_FACTOR
-                        ccReflection.rgb *= litArgs_specularityFactor;
-                    #endif
-
-                    #ifdef LIT_SHEEN
-                        addReflectionSheen(litArgs_worldNormal, dViewDirW, litArgs_sheen_gloss);
-                    #endif
-
-                    // Fresnel has to be applied to reflections
-                    addReflection(dReflDirW, litArgs_gloss);
-
-                    #ifdef LIT_FRESNEL_MODEL
-
-                        dReflection.rgb *= getFresnel(
-                            dot(dViewDirW, litArgs_worldNormal), 
-                            litArgs_gloss, 
-                            litArgs_specularity
-                        #if defined(LIT_IRIDESCENCE)
-                            , iridescenceFresnel,
-                            litArgs_iridescence_intensity
-                        #endif
-                            );
-
-                    #else
-
-                        dReflection.rgb *= litArgs_specularity;
-
-                    #endif
-
-                    #ifdef LIT_SPECULARITY_FACTOR
-                        dReflection.rgb *= litArgs_specularityFactor;
-                    #endif
-
-                #endif
-
-                #ifdef AREA_LIGHTS
-                    // specular has to be accumulated differently if we want area lights to look correct
-                    dSpecularLight *= litArgs_specularity;
-
-                    #ifdef LIT_SPECULAR
-                        // evaluate material based area lights data, shared by all area lights
-                        calcLTCLightValues(litArgs_gloss, litArgs_worldNormal, dViewDirW, litArgs_specularity, litArgs_clearcoat_gloss, litArgs_clearcoat_worldNormal, litArgs_clearcoat_specularity);
-                    #endif
-                #endif
-            `);
-
-            backend.append(`
-                // LOOP - evaluate all non-clustered lights
-                #include "lightEvaluationPS, LIGHT_COUNT"
-            `);
-
-            // clustered lighting
-            if (options.clusteredLightingEnabled && this.lighting) {
-                backend.append(`    addClusteredLights(
-                                        litArgs_worldNormal, 
-                                        dViewDirW, 
-                                        dReflDirW,
-                                #if defined(LIT_CLEARCOAT)
-                                        ccReflDirW,
-                                #endif
-                                        litArgs_gloss, 
-                                        litArgs_specularity, 
-                                        dVertexNormalW, 
-                                        dTBN, 
-                                #if defined(LIT_IRIDESCENCE)
-                                        iridescenceFresnel,
-                                #endif
-                                        litArgs_clearcoat_worldNormal, 
-                                        litArgs_clearcoat_gloss,
-                                        litArgs_sheen_gloss,
-                                        litArgs_iridescence_intensity
-                                    );`);
-            }
-
-            backend.append(`
-
-                #ifdef AREA_LIGHTS
-
-                    #ifdef LIT_CLEARCOAT
-                        // specular has to be accumulated differently if we want area lights to look correct
-                        litArgs_clearcoat_specularity = 1.0;
-                    #endif
-
-                    #ifdef LIT_SPECULAR
-                        litArgs_specularity = vec3(1);
-                    #endif
-
-                #endif
-
-                #ifdef LIT_REFRACTION
-                    addRefraction(
-                        litArgs_worldNormal, 
-                        dViewDirW, 
-                        litArgs_thickness, 
-                        litArgs_gloss, 
-                        litArgs_specularity, 
-                        litArgs_albedo, 
-                        litArgs_transmission,
-                        litArgs_ior,
-                        litArgs_dispersion
-                        #if defined(LIT_IRIDESCENCE)
-                            , iridescenceFresnel, 
-                            litArgs_iridescence_intensity
-                        #endif
-                    );
-                #endif
-            `);
-        }
-
-        backend.append(`
-
-            // apply ambient occlusion
-            #ifdef LIT_AO
-                #ifdef LIT_OCCLUDE_DIRECT
-                    occludeDiffuse(litArgs_ao);
-                #endif
-
-                #if LIT_OCCLUDE_SPECULAR != NONE
-                    occludeSpecular(litArgs_gloss, litArgs_ao, litArgs_worldNormal, dViewDirW);
-                #endif
-            #endif
-
-            #ifdef LIT_SPECULARITY_FACTOR
-                dSpecularLight *= litArgs_specularityFactor;
-            #endif
-
-            #if !defined(LIT_OPACITY_FADES_SPECULAR)
-
-                #if LIT_BLEND_TYPE == NORMAL || LIT_BLEND_TYPE == PREMULTIPLIED
-
-                    float specLum = dot((dSpecularLight + dReflection.rgb * dReflection.a), vec3( 0.2126, 0.7152, 0.0722 ));
-                    #ifdef LIT_CLEARCOAT
-                        specLum += dot(ccSpecularLight * litArgs_clearcoat_specularity + ccReflection.rgb * litArgs_clearcoat_specularity, vec3( 0.2126, 0.7152, 0.0722 ));
-                    #endif
-                    litArgs_opacity = clamp(litArgs_opacity + gammaCorrectInput(specLum), 0.0, 1.0);
-
-                #endif
-
-                litArgs_opacity *= material_alphaFade;
-
-            #endif
-
-            #include "endPS"
-            #include "outputAlphaPS"
-
-            #ifdef LIT_MSDF
-                gl_FragColor = applyMsdf(gl_FragColor);
-            #endif
-
-            #include "outputPS"
-            #include "debugOutputPS"
-        `);
-
-        func.append(`
-            void evaluateBackend() {
-                ${backend.code}
-            }
-        `);
-
-        const result = this._fsGetBeginCode() +
-            this.varyingsCode +
-            this.frontendDecl +
-            decl.code +
-            func.code +
-            code;
-
-        return result;
+        this.fDefineSet(usePerspectiveDepth, 'PERSPECTIVE_DEPTH');
+        this.fDefineSet(true, 'LIGHT_TYPE', `${lightTypeNames[lightType]}`);
+        this.fDefineSet(true, 'SHADOW_TYPE', `${shadowInfo.name}`);
     }
 
     /**
@@ -1086,55 +483,45 @@ class LitShader {
     generateFragmentShader(frontendDecl, frontendCode, lightingUv) {
         const options = this.options;
 
-        this.frontendDecl = frontendDecl;
-        this.frontendCode = frontendCode;
-        this.lightingUv = lightingUv;
-
         if (options.pass === SHADER_PICK || options.pass === SHADER_DEPTH || options.pass === SHADER_PREPASS) {
 
             this.fshader = `
 
-                ${this.device.textureFloatRenderable ? '#define TEXTURE_FLOAT_RENDERABLE' : ''}
-
-                ${this._fsGetBeginCode()}
                 ${this.varyingsCode}
-                ${this.frontendDecl}
-                ${this.frontendCode}
-
-                #ifdef PICK_PASS
-                    #include "pickPS"
-                #endif
-
-                #ifdef PREPASS_PASS
-                    #include "floatAsUintPS"
-                #endif
-
-                void main(void) {
-                    evaluateFrontend();
-
-                    #ifdef PICK_PASS
-                        gl_FragColor = getPickOutput();
-                    #endif
-
-                    #ifdef DEPTH_PASS
-                        gl_FragColor = vec4(1.0, 1.0, 1.0, 1.0);
-                    #endif
-
-                    #ifdef PREPASS_PASS
-                        #if defined(TEXTURE_FLOAT_RENDERABLE)
-                            gl_FragColor = vec4(vLinearDepth, 1.0, 1.0, 1.0);
-                        #else
-                            gl_FragColor = float2uint(vLinearDepth);
-                        #endif
-                    #endif
-                }
+                ${frontendDecl}
+                ${frontendCode}
+                #include "litOtherMainPS"
             `;
-        } else if (this.shadowPass) {
-            this.fshader = this._fsGetShadowPassCode();
-        } else if (options.customFragmentShader) {
-            this.fshader = this._fsGetBeginCode() + options.customFragmentShader;
-        } else {
-            this.fshader = this._fsGetLitPassCode();
+
+        } else if (this.shadowPass) { // SHADOW PASS
+
+            this.prepareShadowPass();
+            this.fshader = `
+                ${this.varyingsCode}
+                ${frontendDecl}
+                ${frontendCode}
+                #include "litShadowMainPS"
+            `;
+
+        } else if (options.customFragmentShader) {   // CUSTOM FRAGMENT SHADER
+
+            this.fshader = `
+                ${options.customFragmentShader}
+            `;
+
+        } else { // FORWARD PASS
+
+            this.prepareForwardPass(lightingUv);
+            this.fshader = `
+                ${this.varyingsCode}
+                ${frontendDecl}
+                #include "litForwardDeclarationPS"
+                #include "litForwardPreCodePS"
+                ${frontendCode}
+                #include "litForwardPostCodePS"
+                #include "litForwardBackendPS"
+                #include "litForwardMainPS"
+            `;
         }
 
         Debug.assert(!this.fshader.includes('litShaderArgs'), 'Automatic compatibility with shaders using litShaderArgs has been removed. Please update the shader to use the new system.');
